# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API provided by DTI project",
        "title": "Domain Threat Intelligence API",
        "contact": {
            "name": "Yaroslav Lysak",
            "url": "https://t.me/Qvineox"
        },
        "version": "0.0.4"
    },
    "paths": {
...
        "/scanning/agents/agent": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts and creates scanning agent",
                "tags": [
                    "Agents"
                ],
                "summary": "Create scanning agent",
                "parameters": [
                    {
                        "description": "agent data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.scanAgentParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts and deletes single scanning agent",
                "tags": [
                    "Agents"
                ],
                "summary": "Delete scanning agent",
                "parameters": [
                    {
                        "description": "agent UUID to delete",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.byUUIDParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts and updates scanning agent",
                "tags": [
                    "Agents"
                ],
                "summary": "Update scanning agent",
                "parameters": [
                    {
                        "description": "agent data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.scanAgentParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/scanning/agents/agent/{agent_uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Return registered agent by UUID",
                "tags": [
                    "Agents"
                ],
                "summary": "Get scanning agent by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent UUID",
                        "name": "agent_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agentEntities.ScanAgent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/scanning/agents/agents": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all registered agents",
                "tags": [
                    "Agents"
                ],
                "summary": "Get all scanning agent",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/agentEntities.ScanAgent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "agentEntities.ScanAgent": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "Description": {
                    "type": "string"
                },
                "Host": {
                    "type": "string"
                },
                "IPAddress": {
                    "$ref": "#/definitions/pgtype.Inet"
                },
                "IsActive": {
                    "type": "boolean"
                },
                "IsConnected": {
                    "description": "IsConnected is used to monitor dialer connection",
                    "type": "boolean"
                },
                "IsHomeBound": {
                    "type": "boolean"
                },
                "IsPrivate": {
                    "description": "Private agents can only be used by their owners.",
                    "type": "boolean"
                },
                "MinPriority": {
                    "description": "MinPriority is minimal job priority that Agent can accept",
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Owner": {
                    "description": "Defines who is the owner of Agent.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/userEntities.PlatformUser"
                        }
                    ]
                },
                "OwnerID": {
                    "type": "integer"
                },
                "UUID": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "config": {
                    "$ref": "#/definitions/datatypes.JSONType-agentEntities_ScanAgentConfig"
                }
            }
        },
...
        "datatypes.JSONType-agentEntities_ScanAgentConfig": {
            "type": "object"
        },
...
        "error.APIError": {
            "type": "object",
            "properties": {
                "ErrorCode": {
                    "$ref": "#/definitions/error.CodeError"
                },
                "ErrorMessage": {
                    "type": "string"
                },
                "ErrorModule": {
                    "type": "string"
                },
                "StatusCode": {
                    "type": "integer"
                }
            }
        },
        "error.CodeError": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15
            ],
            "x-enum-varnames": [
                "EncodingErrorCode",
                "DecodingErrorCode",
                "IncorrectParamsErrorCode",
                "InsufficientParamsErrorCode",
                "DatabaseErrorCode",
                "DatabaseEntityNotFoundCode",
                "FileDecodingErrorCode",
                "FileReadingErrorCode",
                "FileExtensionNotSupportedErrorCode",
                "NotImplementedErrorCode",
                "FileProcessingErrorCode",
                "InternalUnidentifiedErrorCode",
                "AuthFailedErrorCode",
                "AuthPermissionInsufficientErrorCode",
                "JobQueueErrorCode"
            ]
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "jobEntities.JobPriority": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-comments": {
                "JOB_PRIORITY_CRITICAL": "job must be executed instantly",
                "JOB_PRIORITY_HIGH": "job must be executed after current (stack mode)",
                "JOB_PRIORITY_LOW": "job should be executed lastly in order (queue mode)",
                "JOB_PRIORITY_MEDIUM": "job should be executed with higher priority"
            },
            "x-enum-varnames": [
                "JOB_PRIORITY_CRITICAL",
                "JOB_PRIORITY_HIGH",
                "JOB_PRIORITY_MEDIUM",
                "JOB_PRIORITY_LOW"
            ]
        },
...
        "net.IPNet": {
            "type": "object",
            "properties": {
                "ip": {
                    "description": "network number",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "mask": {
                    "description": "network mask",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "pgtype.Inet": {
            "type": "object",
            "properties": {
                "ipnet": {
                    "$ref": "#/definitions/net.IPNet"
                },
                "status": {
                    "$ref": "#/definitions/pgtype.Status"
                }
            }
        },
        "pgtype.Status": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "Undefined",
                "Null",
                "Present"
            ]
        },
...
        "routing.byUUIDParams": {
            "type": "object",
            "required": [
                "UUID"
            ],
            "properties": {
                "UUID": {
                    "type": "string"
                }
            }
        },
...
        "routing.scanAgentParams": {
            "type": "object",
            "required": [
                "Host",
                "Name"
            ],
            "properties": {
                "Description": {
                    "type": "string"
                },
                "Host": {
                    "type": "string"
                },
                "IsActive": {
                    "type": "boolean"
                },
                "IsHomeBound": {
                    "type": "boolean"
                },
                "IsPrivate": {
                    "type": "boolean"
                },
                "MinPriority": {
                    "$ref": "#/definitions/jobEntities.JobPriority"
                },
                "Name": {
                    "type": "string"
                },
                "OwnerID": {
                    "type": "integer"
                },
                "UUID": {
                    "type": "string"
                }
            }
        },
...
        "userEntities.PlatformUser": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "Email": {
                    "type": "string"
                },
                "FullName": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "IsActive": {
                    "type": "boolean"
                },
                "Login": {
                    "type": "string"
                },
                "Permissions": {
                    "description": "Defines which roles user has",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userEntities.PlatformUserPermission"
                    }
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
        "userEntities.PlatformUserPermission": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "Description": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "IsActive": {
                    "type": "boolean"
                },
                "Name": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-Key",
            "in": "header"
        }
    }
}